feat(dart_frog_cli): add devserver domain
