+++ packages/dart_frog_cli/lib/src/daemon/daemon.dart
@@ -1,5 +1,5 @@
 export 'connection.dart';
 export 'daemon_server.dart';
-export 'domain.dart';
+export 'domain/domain.dart';
 export 'exceptions.dart';
 export 'protocol.dart';
